Adjust/separate base controller for plugin + route prefix
