+++ Comments/src/Controller/Admin/Controller.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Croogo\Comments\Controller\Admin;
+
+use Croogo\Core\Controller\Admin\Controller as CroogoController;
+
+/**
+ * Comments Admin Controller
+ *
+ * @category Comments.Controller
+ * @package  Croogo.Comments.Controller
+ * @version  1.0
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+}
