+++ Contacts/src/Controller/Admin/ContactsController.php
@@ -3,7 +3,7 @@
 namespace Croogo\Contacts\Controller\Admin;
 
 use App\Network\Email\Email;
-use Croogo\Contacts\Controller\ContactsAppController;
+use Croogo\Contacts\Controller\Admin\Controller;
 use Croogo\Contacts\Model\Entity\Contact;
 
 /**
@@ -16,7 +16,7 @@ use Croogo\Contacts\Model\Entity\Contact;
  * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
  * @link     http://www.croogo.org
  */
-class ContactsController extends ContactsAppController {
+class ContactsController extends Controller {
 
 /**
  * Components
