+++ Contacts/src/Controller/Admin/Controller.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Croogo\Contacts\Controller\Admin;
+
+use Croogo\Core\Controller\Admin\Controller as CroogoController;
+
+/**
+ * Contacts Admin Controller
+ *
+ * @category Contacts.Controller
+ * @package  Croogo.Contacts.Controller
+ * @version  1.0
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+}
