+++ Contacts/src/Controller/Controller.php
@@ -2,7 +2,8 @@
 
 namespace Croogo\Contacts\Controller;
 
-use Croogo\Core\Controller\CroogoAppController;
+use Croogo\Core\Controller\Controller as CroogoController;
+
 /**
  * Contacts App Controller
  *
@@ -13,6 +14,6 @@ use Croogo\Core\Controller\CroogoAppController;
  * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
  * @link     http://www.croogo.org
  */
-class ContactsAppController extends CroogoAppController {
+class Controller extends CroogoController {
 
 }
