+++ Extensions/src/Controller/Admin/DashboardController.php
@@ -3,7 +3,8 @@
 namespace Croogo\Extensions\Controller\Admin;
 
 use Croogo\Extensions\Controller\Admin\Controller as CroogoController;
-class ExtensionsDashboardController extends ExtensionsAppController {
+
+class DashboardController extends CroogoController {
 
 /**
  * Admin dashboard
