+++ FileManager/src/Controller/Admin/Controller.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Croogo\FileManager\Controller\Admin;
+
+use Croogo\Core\Controller\Admin\Controller as CroogoController;
+
+/**
+ * FileManager Admin Controller
+ *
+ * @category FileManager.Controller
+ * @package  Croogo.FileManager.Controller
+ * @version  1.4
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+}
