+++ Menus/src/Controller/Admin/LinksController.php
@@ -5,9 +5,8 @@ namespace Croogo\Menus\Controller\Admin;
 use Cake\Core\Configure;
 use Cake\Event\Event;
 use Croogo\Core\Controller\Component\CroogoComponent;
-use Croogo\Core\Controller\Admin\Controller;
 use Croogo\Core\Croogo;
-use Croogo\Menus\Controller\MenusAppController;
+use Croogo\Menus\Controller\Admin\Controller;
 use Croogo\Menus\Model\Table\LinksTable;
 
 /**
