+++ Nodes/src/Controller/Admin/Controller.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Croogo\Nodes\Controller\Admin;
+
+use Croogo\Core\Controller\Admin\Controller as CroogoController;
+
+/**
+ * Nodes Admin Controller
+ *
+ * @category Nodes.Controller
+ * @package  Croogo.Nodes
+ * @since  1.5
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+}
