+++ Taxonomy/src/Controller/Controller.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Croogo\Taxonomy\Controller;
+
+use Croogo\Core\Controller\Controller as CroogoController;
+
+/**
+ * Taxonomy App Controller
+ *
+ * @category Taxonomy.Controller
+ * @package  Croogo
+ * @version  1.4
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+}
