+++ Users/src/Controller/Admin/Controller.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Croogo\Users\Controller\Admin;
+
+use Croogo\Core\Controller\Admin\Controller as CroogoController;
+
+/**
+ * Users Admin Controller
+ *
+ * @category Controllers
+ * @package  Croogo.Users.Controller
+ * @since    1.5
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+
+}
