+++ Users/src/Controller/Controller.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Croogo\Users\Controller;
+
+use Croogo\Core\Controller\Controller as CroogoController;
+
+/**
+ * Users Application controller
+ *
+ * @category Controllers
+ * @package  Croogo.Users.Controller
+ * @since    1.5
+ * @author   Fahad Ibnay Heylaal <contact@fahad19.com>
+ * @license  http://www.opensource.org/licenses/mit-license.php The MIT License
+ * @link     http://www.croogo.org
+ */
+class Controller extends CroogoController {
+
+
+}
